+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/metadata/MetadataContextHolderTest.java
@@ -13,6 +13,7 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
+ *
  */
 
 package com.tencent.cloud.common.metadata;
@@ -45,10 +46,10 @@ public class MetadataContextHolderTest {
 		customMetadata.put("a", "1");
 		customMetadata.put("b", "2");
 		MetadataContext metadataContext = MetadataContextHolder.get();
-		metadataContext.putAllTransitiveCustomMetadata(customMetadata);
+		metadataContext.putFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE, customMetadata);
 		MetadataContextHolder.set(metadataContext);
 
-		customMetadata = MetadataContextHolder.get().getAllTransitiveCustomMetadata();
+		customMetadata = MetadataContextHolder.get().getFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE);
 		Assertions.assertThat(customMetadata.get("a")).isEqualTo("1");
 		Assertions.assertThat(customMetadata.get("b")).isEqualTo("2");
 
@@ -60,7 +61,7 @@ public class MetadataContextHolderTest {
 		customMetadata.put("c", "3");
 		MetadataContextHolder.init(customMetadata);
 		metadataContext = MetadataContextHolder.get();
-		customMetadata = metadataContext.getAllTransitiveCustomMetadata();
+		customMetadata = metadataContext.getFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE);
 		Assertions.assertThat(customMetadata.get("a")).isEqualTo("1");
 		Assertions.assertThat(customMetadata.get("b")).isEqualTo("22");
 		Assertions.assertThat(customMetadata.get("c")).isEqualTo("3");