+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change Log
 ---
 
+- [fix:Router support request label.](https://github.com/Tencent/spring-cloud-tencent/pull/165)
+